Base project sans erreur.
