--- app/src/main/java/com/example/multuscalendrius/LoginResponse.java
@@ -1,13 +1,0 @@
-package com.example.multuscalendrius;
-
-public class LoginResponse {
-    private Boolean token;
-
-    public Boolean getToken() {
-        return token;
-    }
-
-    public void setToken(Boolean token) {
-        this.token = token;
-    }
-}
