+++ app/src/main/java/com/example/multuscalendrius/modeles/entitees/UserRole.java
@@ -1,0 +1,7 @@
+package com.example.multuscalendrius.modeles.entitees;
+
+public enum UserRole {
+    AUTHOR,  // Possède tous les droits sur le calendrier
+    MEMBER,  // Peut créer, modifier et supprimer des événements
+    GUEST    // Accès en lecture seule
+}
