--- app/src/main/java/com/example/multuscalendrius/objects/Calendar.java
@@ -1,55 +1,0 @@
-package com.example.multuscalendrius.objects;
-import java.util.Date;
-
-public class Calendar {
-    private Long id;
-    private String name;
-    private String description;
-    private Date createdAt;
-    private Date updatedAt;
-
-    // Constructeur par défaut
-    public Calendar() {}
-
-    // Constructeur complet
-    public Calendar(Long id, String name, String description, Date createdAt, Date updatedAt) {
-        this.id = id;
-        this.name = name;
-        this.description = description;
-        this.createdAt = createdAt;
-        this.updatedAt = updatedAt;
-    }
-
-    // Getters et Setters
-    public Long getId() {
-        return id;
-    }
-    public void setId(Long id) {
-        this.id = id;
-    }
-    public String getName() {
-        return name;
-    }
-    public void setName(String name) {
-        this.name = name;
-    }
-    public String getDescription() {
-        return description;
-    }
-    public void setDescription(String description) {
-        this.description = description;
-    }
-    public Date getCreatedAt() {
-        return createdAt;
-    }
-    public void setCreatedAt(Date createdAt) {
-        this.createdAt = createdAt;
-    }
-    public Date getUpdatedAt() {
-        return updatedAt;
-    }
-    public void setUpdatedAt(Date updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-}
-
