--- app/src/main/java/com/example/multuscalendrius/objects/User.java
@@ -1,45 +1,0 @@
-package com.example.multuscalendrius.objects;
-
-public class User {
-    private Long id;
-    private String email;
-    private String password;
-    private String name;
-
-    public User() {}
-
-    // Constructeur complet
-    public User(Long id, String email, String password, String name) {
-        this.id = id;
-        this.email = email;
-        this.password = password;
-        this.name = name;
-    }
-
-    // Getters et Setters
-    public Long getId() {
-        return id;
-    }
-    public void setId(Long id) {
-        this.id = id;
-    }
-    public String getEmail() {
-        return email;
-    }
-    public void setEmail(String email) {
-        this.email = email;
-    }
-    public String getPassword() {
-        return password;
-    }
-    public void setPassword(String password) {
-        this.password = password;
-    }
-    public String getName() {
-        return name;
-    }
-    public void setName(String name) {
-        this.name = name;
-    }
-}
-
