--- app/src/main/java/com/example/multuscalendrius/objects/UserRole.java
@@ -1,7 +1,0 @@
-package com.example.multuscalendrius.objects;
-
-public enum UserRole {
-    AUTHOR,  // Possède tous les droits sur le calendrier
-    MEMBER,  // Peut créer, modifier et supprimer des événements
-    GUEST    // Accès en lecture seule
-}
