+++ app/src/main/java/com/example/multuscalendrius/vuemodele/VueModele.java
@@ -1,0 +1,4 @@
+package com.example.multuscalendrius.vuemodele;
+
+public class VueModele {
+}
