+++ app/src/main/java/com/example/multuscalendrius/vues/LoginActivity.java
@@ -1,34 +1,23 @@
-package com.example.multuscalendrius;
+package com.example.multuscalendrius.vues;
 
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.TextView;
-import android.widget.Toast;
+import android.widget.*;
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.multuscalendrius.R;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import org.json.JSONObject;
 
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-import okhttp3.Call;
-import okhttp3.Callback;
-import okhttp3.MediaType;
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.RequestBody;
-import okhttp3.Response;
+import okhttp3.*;
 
 public class LoginActivity extends AppCompatActivity {
 
     private EditText editTextEmail, editTextPassword;
     private Button buttonLogin;
-    private Button test;
     private TextView textViewSignUp;
     private OkHttpClient client;
     private JSONObject obj;
@@ -51,7 +40,6 @@ public class LoginActivity extends AppCompatActivity {
 
         // Action du bouton de connexion
         buttonLogin.setOnClickListener(v -> login());
-        test.setOnClickListener(v -> login());
 
         // Action pour accéder à la page d'inscription
         textViewSignUp.setOnClickListener(v -> {
@@ -93,23 +81,24 @@ public class LoginActivity extends AppCompatActivity {
 
         try(Response response = client.newCall(request).execute()){
             if (response.isSuccessful()) {
-                String responseBody = response.body().string();
-                try {
-                    // Supposons que l'API renvoie un objet JSON avec un champ "token"
-                    LoginResponse loginResponse = mapper.readValue(responseBody, LoginResponse.class);
-                    runOnUiThread(() -> {
+                if (response.body() != null) {
+                    String responseBody = response.body().string();
+                    try {
+                        // Supposons que l'API renvoie un objet JSON avec un champ "token"
+                        LoginResponse loginResponse = mapper.readValue(responseBody, LoginResponse.class);
+                        runOnUiThread(() -> {
 
-                        // Rediriger vers le Dashboard (ou l'écran principal)
-                        Intent intent = new Intent(LoginActivity.this, DashboardActivity.class);
-                        startActivity(intent);
-                        finish();
-                    });
-                } catch (JsonProcessingException e) {
-                    e.printStackTrace();
-                    runOnUiThread(() -> Toast.makeText(LoginActivity.this, "Erreur de parsing", Toast.LENGTH_SHORT).show());
+                            // Rediriger vers le Dashboard (ou l'écran principal)
+                            //Intent intent = new Intent(LoginActivity.this, DashboardActivity.class);
+                            //startActivity(intent);
+                        });
+                    } catch (JsonProcessingException e) {
+                        e.printStackTrace();
+                        runOnUiThread(() -> Toast.makeText(LoginActivity.this, "Erreur de parsing", Toast.LENGTH_SHORT).show());
+                    }
+                } else {
+                    runOnUiThread(() -> Toast.makeText(LoginActivity.this, "Échec de la connexion", Toast.LENGTH_SHORT).show());
                 }
-            } else {
-                runOnUiThread(() -> Toast.makeText(LoginActivity.this, "Échec de la connexion", Toast.LENGTH_SHORT).show());
             }
         }
         catch (IOException e){
