+++ app/src/main/java/com/example/multuscalendrius/vues/LoginResponse.java
@@ -1,0 +1,13 @@
+package com.example.multuscalendrius.vues;
+
+public class LoginResponse {
+    private Boolean token;
+
+    public Boolean getToken() {
+        return token;
+    }
+
+    public void setToken(Boolean token) {
+        this.token = token;
+    }
+}
