+++ app/src/main/java/com/example/multuscalendrius/vues/SignUpActivity.java
@@ -1,4 +1,4 @@
-package com.example.multuscalendrius;
+package com.example.multuscalendrius.vues;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,6 +7,8 @@ import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.Toast;
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.multuscalendrius.R;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
