+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,9 @@
 <resources>
     <string name="app_name">multus calendrius</string>
+    <string name="inscrire">S\'inscrire</string>
+    <string name="deja_inscrit_se_connecter">Déjà inscrit ? Se connecter</string>
+    <string name="confirmer_le_mot_de_passe">Confirmer le mot de passe</string>
+    <string name="mot_de_passe">Mot de passe</string>
+    <string name="adresse_courriel">Adresse courriel</string>
+    <string name="se_connecter">Se connecter</string>
 </resources>
